Add `BuilderContext` interface to encapsulate `RunContext` and provide additional builder constructor requirements.
